MAINT: renamed io.lammps_data to io.lammpsdata ASE consistency
